Add flags and restructure tile hash, tile and cover functions
