+++ config/config.go
@@ -139,8 +139,13 @@ type _ImportOptions struct {
 	DiffStateBefore  time.Duration
 }
 
+type _DiffOptions struct {
+	TileList string
+}
+
 var BaseOptions = _BaseOptions{}
 var ImportOptions = _ImportOptions{}
+var DiffOptions = _DiffOptions{}
 
 func addBaseFlags(flags *flag.FlagSet) {
 	flags.StringVar(&BaseOptions.Connection, "connection", "", "connection parameters")
@@ -175,6 +180,9 @@ func init() {
 	DiffFlags.Usage = UsageDiff
 
 	addBaseFlags(DiffFlags)
+	//TODO: What is a smart default here? Just none?
+	DiffFlags.StringVar(&DiffOptions.TileList, "tilelist", "", "expired tiles")
+
 	addBaseFlags(ImportFlags)
 	ImportFlags.BoolVar(&ImportOptions.Overwritecache, "overwritecache", false, "overwritecache")
 	ImportFlags.BoolVar(&ImportOptions.Appendcache, "appendcache", false, "append cache")
